Adds first pass OBJ loader, fixes some warnings

+++ src/renderer/buffer.rs
@@ -102,7 +102,7 @@ impl InternalBuffer {
         if let Some(allocation) = &self.allocation {
             let data_ptr = allocation.mapped_ptr().unwrap().as_ptr() as *mut u8;
             unsafe {
-                data_ptr.copy_from_nonoverlapping(data.as_ptr() as *const u8, data_len as usize)
+                data_ptr.copy_from_nonoverlapping(data.as_ptr() as *const u8, data_len)
             };
         } else {
             panic!("Buffer had no allocation!");
